Split out the main view handlers a little more. Starting work on matrix handlers.
