+++ frontend/package.json
@@ -20,13 +20,11 @@
 	"dependencies": {
 		"@datapunt/matomo-tracker-react": "0.3.1",
 		"@date-io/date-fns": "1.3.13",
-		"@dod-advana/advana-jira-feedback": "0.4.15",
 		"@dod-advana/advana-platform-ui": "1.4.4",
 		"@dod-advana/advana-side-nav": "0.0.17",
 		"@dod-advana/advana-tutorial-overlay": "0.0.5",
 		"@emotion/react": "11.7.1",
 		"@emotion/styled": "11.6.0",
-		"@marvelapp/react-ab-test": "3.1.0",
 		"@material-ui/core": "4.11.0",
 		"@material-ui/icons": "4.9.1",
 		"@material-ui/lab": "4.0.0-alpha.55",
@@ -66,7 +64,7 @@
 		"react-collapse": "4.0.3",
 		"react-dom": "17.0.2",
 		"react-error-boundary": "3.1.4",
-		"react-force-graph": "1.41.7",
+		"react-force-graph-2d": "^1.23.10",
 		"react-js-pagination": "3.0.3",
 		"react-loadable": "5.5.0",
 		"react-loadable-visibility": "3.0.2",
