+++ frontend/src/components/mainView/commonStyles.js
@@ -1,0 +1,170 @@
+export const fullWidthCentered = {
+	width: '100%',
+	display: 'flex',
+	flexDirection: 'column',
+	justifyContent: 'center',
+	alignItems: 'center',
+};
+
+export const styles = {
+	listViewBtn: {
+		minWidth: 0,
+		margin: '20px 0px 0px',
+		marginLeft: 10,
+		padding: '0px 7px 0',
+		fontSize: 20,
+		height: 34,
+	},
+	cachedResultIcon: {
+		display: 'flex',
+		justifyContent: 'center',
+		padding: '0 0 1% 0',
+	},
+	searchResults: fullWidthCentered,
+	paginationWrapper: fullWidthCentered,
+	tabContainer: {
+		alignItems: 'center',
+		marginBottom: '14px',
+		height: '600px',
+		margin: '0px 4% 0 65px',
+	},
+	tabButtonContainer: {
+		width: '100%',
+		padding: '0em 1em',
+		alignItems: 'center',
+	},
+	spacer: {
+		flex: '0.375',
+	},
+	resultsCount: {
+		fontFamily: 'Noto Sans',
+		fontSize: 22,
+		fontWeight: 'bold',
+		color: '#131E43',
+		paddingTop: '10px',
+	},
+	subtext: {
+		color: '#8091A5',
+		fontSize: 12,
+	},
+	containerText: {
+		fontSize: 16,
+		fontWeight: 'bold',
+	},
+	checkboxPill: {
+		textAlign: 'center',
+		borderRadius: '10px',
+		paddingLeft: '10px',
+		paddingRight: '10px',
+		lineHeight: 1.2,
+		fontSize: '12px',
+		marginLeft: '10px',
+		border: '2px solid #bdccde',
+		backgroundColor: 'white',
+		boxSizing: 'border-box',
+		color: 'black',
+		minHeight: '35px',
+		display: 'flex',
+		alignItems: 'center',
+		justifyContent: 'center',
+		cursor: 'pointer',
+	},
+	leftContainerSummary: {
+		width: '15%',
+		marginTop: 10,
+	},
+	rightContainerSummary: {
+		marginLeft: '17.5%',
+		width: '79.7%',
+	},
+	showingResultsRow: {
+		width: '100%',
+		display: 'inline-flex',
+		justifyContent: 'space-between',
+		marginBottom: 10,
+		marginTop: 15,
+	},
+	container: {
+		minWidth: 148,
+	},
+	containerDiv: {
+		marginTop: 10,
+		marginLeft: 0,
+		marginRight: 0,
+	},
+	listViewContainer: {
+		marginTop: 10,
+		marginLeft: 0,
+		marginRight: 0,
+		paddingRight: 40,
+	},
+	text: {
+		margin: 'auto 0px',
+	},
+	buttons: {
+		height: 50,
+		width: 64,
+		margin: '0px 5px',
+		minWidth: 64,
+	},
+	unselectedButton: {
+		border: 'solid 2px #DFE6EE',
+		color: '#8091A5',
+	},
+	icon: {
+		fontSize: 30,
+	},
+	filterBox: {
+		backgroundColor: '#ffffff',
+		borderRadius: '5px',
+		padding: '2px',
+		border: '2px solid #bdccde',
+		pointerEvents: 'none',
+		margin: '2px 5px 0px',
+	},
+	titleText: {
+		fontSize: 22,
+		fontWeight: 500,
+		color: '#131E43',
+		margin: '20px 0',
+		fontFamily: 'Montserrat',
+	},
+	tableColumn: {
+		textAlign: 'center',
+		margin: '4px 0',
+	},
+	tabsList: {
+		borderBottom: `2px solid ${'#1C2D65'}`,
+		padding: 0,
+		display: 'flex',
+		alignItems: 'center',
+		flex: 9,
+		marginRight: 15,
+	},
+	tabStyle: {
+		border: `1px solid ${'#DCDCDC'}`,
+		borderBottom: 'none !important',
+		borderRadius: `6px 6px 0px 0px`,
+		position: ' relative',
+		listStyle: 'none',
+		padding: '2px 12px',
+		cursor: 'pointer',
+		textAlign: 'center',
+		backgroundColor: '#ffffff',
+		marginRight: '2px',
+		marginLeft: '2px',
+		height: 45,
+		display: 'flex',
+		alignItems: 'center',
+		justifyContent: 'center',
+	},
+	tabSelectedStyle: {
+		border: `1px solid ${'#0000001F'}`,
+		backgroundColor: '#1C2D65',
+		color: 'white',
+	},
+	orangeText: {
+		fontWeight: 'bold',
+		color: '#E9691D',
+	},
+};
